Separated client-sided fileInput
