+++ frontend/src/components/FileInput/FileInputEditor/index.css
@@ -1,0 +1,29 @@
+.upload-file-container {
+  display: flex;
+  width: 100%;
+  justify-content: center;
+  align-items: center;
+  padding-left: 5px;
+  padding-right: 5px;
+  height: 200px;
+}
+
+.upload-file-container-inner {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  justify-content: center;
+  height: 100%;
+}
+
+.upload-icon {
+  height: 22px;
+  width: 22px;
+  object-fit: contain;
+  margin: 10px;
+}
+
+.upload-text {
+  font-size: 16px;
+  color: #666666;
+}
